Introduce canvas accessors and sizexy.
All turtles are now <canvas> instead of <img> by default.

The global function canvas() provides access to the fullpage canvas
that the turtles draw on by default.

The turtle method turtle.canvas() provides acccess to the per-turtle
canvas.

The sizexy() global method provides access to the current dimensions
of the document.

And wear('100x200') or wear('red 100x200/2') generates a blank canvas of
dimensions 100x200, or a red 100x200 canvas shown half size on screen.