Add interrupt support.
Adds the interrupt() function that aborts pending queued animations,
throws an exception, and causes every turtle motion to also throw
an exception.  It can be used to stop a program and exit an infinite
loop.

Also adds support for auto-interrupt of infinite loops that contain
turtle motions.  After 100 turtle motions, a zero-timeout it set up,
and if more than 6 seconds passes without processing the timeout, the
browser is considered hung, and an interrupt() is called.