Add Sprite as a base class of Turtle.
This commit rationalizes the class hierarchy of Turtle a bit.

The default shape for a Sprite is a blank 256x256 rectangle
instead of a turtle shape; otherwise it's the same as a Turtle.

A Sprite can also be constructed with an object like this:

s = new Sprite { width: 10, height: 50, color: red }