Add menu and type for ASCII-print-goto-programs.
These two functions enable an ASCII-art-and-goto-style of programming
(such as "make your own adventure") that is educational and also easy
for beginners to understand.

type prints raw characters to the screen without extra formatting,
by appending them into the last <PRE> in the body (creating one if
one isn't there).

menu shows a one-shot radio button menu, and can invoke one of
a selection of callbacks as soon as the user clicks.  This
radio button menu has been carefully tested for keyboard
accessibility on IE, Chrome, and FF.

To make a make-your-own adventure, make a function for each "page".
Each function types out the page, and then shows a menu
lists all the possible successor pages (and their functions)
in a map.