Draft of instrument support.
This adds an instrument class for audio support along with a
syntax for specifying timbre.

An instrument is basically a tone generator that can also do
sequencing.  It is possible to make a immediate-mode piano using the
turtle.tone(freq) command.  turtle.tone(freq, false) stops the sound.
The old method turtle.play(song) is still there, and it now works
by schedules a bunch of tone() over time.

An instrument cannot play more than one tone with the same frequency
(it will end the earlier one when another one starts), and it also
only sends a few seconds of tones to WebAudio at a time.  This
architecture keeps the load on WebAudio light, and helps solve the
"too many oscillators" problem.